indexer/requirements.txt
awslambdaric
boto3
cryptography
openai
pdfminer.six
protobuf
psycopg2-binary
sentence-transformers
slack-bolt
SQLAlchemy
sqlalchemy-utils
torch